% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{a_test_results}
\alias{a_test_results}
\title{Analysed results from tutorial_consistency_set: a-test scores when sets compared}
\format{A list with 95 rows and 6 columns}
\usage{
data(a_test_results)
}
\description{
This dataset contains the results of the analysis aa_getATestResults when
applied to the tutorial_consistency_dataset. Used in testing
}
\details{
\itemize{
  \item Sample.Size Sample size being analysed
  \item Sample The number of the subset of this sample size
  \item ATestVelocity A-Test score for the velocity measure
  \item ATestVelocityNorm Normalised A-Test score for the velocity measure
  \item ATestDisplacement A-Test score for the displacement measure
  \item ATestDisplacementNorm Normalised A-Test score for the displacement measure
}
}
\keyword{datasets}
