% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aleatory_analysis_plotting.R
\name{aa_graphATestsForSampleSize}
\alias{aa_graphATestsForSampleSize}
\title{Produce a plot for each sample size, showing the A-Test scores for each set of that size}
\usage{
aa_graphATestsForSampleSize(FILEPATH, ATESTS, MEASURES, LARGEDIFFINDICATOR,
  GRAPHOUTPUTNAME, TIMEPOINT, TIMEPOINTSCALE)
}
\arguments{
\item{FILEPATH}{Directory where the results of the simulation runs, in folders or in single CSV file format, can be found}

\item{ATESTS}{Name of the file where the calculated A-Test scores can be found}

\item{MEASURES}{An array containing the names of the simulation output measures to be analysed.}

\item{LARGEDIFFINDICATOR}{The A-Test determines there is a large difference between two sets if the result is greater than 0.2 either side of the 0.5 line.  Should this not be suitable, this can be changed here}

\item{GRAPHOUTPUTNAME}{Name of the graph to be output for each sample size. Should be in PDF format}

\item{TIMEPOINT}{Timepoint for which this plot is being created}

\item{TIMEPOINTSCALE}{Implemented so this method can be used when analysing multiple simulation timepoints. Sets the scale of the timepoints being analysed, e.g. "Hours"}
}
\description{
Produce a plot for each sample size, showing the A-Test scores for each set of that size
}
