% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netlogo_sampling_utilities.R
\name{add_parameter_value_to_file}
\alias{add_parameter_value_to_file}
\title{Iterates through the parameters, adding their sampled value to the netlogo experiment file}
\usage{
add_parameter_value_to_file(xml, PARAMETERS, ParameterInfo, LHC_DESIGN,
  SAMPLE, PARAMVALS)
}
\arguments{
\item{xml}{Object of the XML file being constructed}

\item{PARAMETERS}{Parameters specified in the R set-up file, of interest in this experiment}

\item{ParameterInfo}{Parameters that are being perturbed, and their ranges}

\item{LHC_DESIGN}{The LHC sample generated for these parameters}

\item{SAMPLE}{Number of the sample being processed}

\item{PARAMVALS}{Input of parameter ranges or set values specified by the user}
}
\value{
Updated XML object ready for output to file
}
\description{
Iterates through the parameters, adding their sampled value to the netlogo experiment file
}
