% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_analysis.R
\name{all_prccs_for_parameter}
\alias{all_prccs_for_parameter}
\title{Calculates PRCC values for all measures for a given parameter}
\usage{
all_prccs_for_parameter(parameter, lhc_results, measures,
  cor_calc_method = c("s"))
}
\arguments{
\item{parameter}{Name of simulation parameter}

\item{lhc_results}{Summary of simulation results for all parameter conditions, with parameter values}

\item{measures}{Simulation output responses}

\item{cor_calc_method}{Way to calculate the correlation coefficient: Pearson's
("p"), Spearman's ("s"), and Kendall's ("k"). Default is p}
}
\value{
PRCC values for all measures for this parameter
}
\description{
Calculates PRCC values for all measures for a given parameter
}
