% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network_utilities.R
\name{analysenetwork_structures}
\alias{analysenetwork_structures}
\title{Analyse each network structure provided as a potential NN structure}
\usage{
analysenetwork_structures(fold_size, dataset, model_formula, parameters,
  measures, algorithm_settings)
}
\arguments{
\item{fold_size}{Number of rows in which the dataset is divided into folds}

\item{dataset}{Dataset used to assess each strutcure}

\item{model_formula}{Parameters and measures formula used to specify the
input and output layers of the hidden network}

\item{parameters}{Names of the parameters that form the input nodes of
the neural network}

\item{measures}{Names of the simulation responses that form the output
node of the neural network}

\item{algorithm_settings}{Object output from the function
emulation_algorithm_settings, containing the settings of the machine
learning algorithms to use in emulation creation. In this case, the
settings parameter we are interested in is number of generations}
}
\value{
mean squared errors for the assessed structures
}
\description{
Analyse each network structure provided as a potential NN structure
}
