% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aleatory_analysis.R
\name{append_time_to_argument}
\alias{append_time_to_argument}
\title{Appends the time to an argument if processing timepoints}
\usage{
append_time_to_argument(argument, current_time, file_format)
}
\arguments{
\item{argument}{Argument to append the current time to}

\item{current_time}{Current time being processed}

\item{file_format}{Format of the file name being altered}
}
\value{
Amended file name argument with time point
}
\description{
Appends the time to an argument if processing timepoints
}
