% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atest_utilities.R
\name{atest}
\alias{atest}
\title{Calculates the A-test score for two distributions}
\usage{
atest(x, y)
}
\arguments{
\item{x, y}{distributions of simulation results}
}
\value{
A-Test score
}
\description{
Calculates the A-test score for two distributions
}
