% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast_analysis.R
\name{build_curve_results_from_r_object}
\alias{build_curve_results_from_r_object}
\title{When developing spartanDB, it became clear curve results may not be in separate
files but in one R object. This takes an R object containing curve summaries and
builds these into the format spartan requires to perform the analysis}
\usage{
build_curve_results_from_r_object(CURVE_SUMMARY, NUMCURVES, NUMSAMPLES,
  MEASURES, PARAMETERS)
}
\arguments{
\item{CURVE_SUMMARY}{R object containing summaries for all curves}

\item{NUMCURVES}{The number of 'resamples' to perform (see eFAST
documentation) - recommend using at least 3}

\item{NUMSAMPLES}{The number of parameter subsets that were generated in
the eFAST design}

\item{MEASURES}{Array containing the names of the output measures which
are used to analyse the simulation}

\item{PARAMETERS}{Array containing the names of the parameters of which
parameter samples have been generated}
}
\value{
R matrix containing the curve summaries in the format spartan requires
}
\description{
When developing spartanDB, it became clear curve results may not be in separate
files but in one R object. This takes an R object containing curve summaries and
builds these into the format spartan requires to perform the analysis
}
