% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulation_utilities.R
\name{build_performance_statistics}
\alias{build_performance_statistics}
\title{Internal function to build performance statistics (MSE, R2, generation
time) for an emulation}
\usage{
build_performance_statistics(technique, model_predictions, observed_data,
  measures, benchmark)
}
\arguments{
\item{technique}{The machine learning technique used to develop the emulator}

\item{model_predictions}{Predicted dataset}

\item{observed_data}{Observed dataset (testing or validation)}

\item{measures}{Simulation responses the model should predict}

\item{benchmark}{Dataset of benchmarking data to which these results are
being added}
}
\value{
Updated benchmark object, probably of numerous machine learning
techniques, showing the MSE, R2, and generation time for this technique
}
\description{
Internal function to build performance statistics (MSE, R2, generation
time) for an emulation
}
\keyword{internal}
