% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness_analysis.R
\name{calculate_atest_score}
\alias{calculate_atest_score}
\title{Calculate the A-Test score for a parameter set in comparison with baseline}
\usage{
calculate_atest_score(parameter_result, exp_params, baseline_result,
  measures, parameter, value)
}
\arguments{
\item{parameter_result}{Simulation results under current parameter set}

\item{exp_params}{Parameter set that led to this behaviour}

\item{baseline_result}{Results obtained at baseline conditions}

\item{measures}{Simulation response measures}

\item{parameter}{Name of the parameter being processed}

\item{value}{Value of the parameter being processsed}
}
\value{
A-Test scores, or NULL if result not found
}
\description{
Calculate the A-Test score for a parameter set in comparison with baseline
}
