% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network_utilities.R
\name{calculate_fold_MSE}
\alias{calculate_fold_MSE}
\title{Calculate the mean squared error for this fold in k-fold cross validation}
\usage{
calculate_fold_MSE(nn_predictions, test_fold, nn, measures)
}
\arguments{
\item{nn_predictions}{Predictions made by neural net}

\item{test_fold}{The test data on which performance is being assessed}

\item{nn}{The generated neural network}

\item{measures}{The simulation output responses that are being predicted}
}
\value{
Mean Squared error for this fold.
}
\description{
Calculate the mean squared error for this fold in k-fold cross validation
}
