% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_analysis.R
\name{calculate_prcc_for_all_measures}
\alias{calculate_prcc_for_all_measures}
\title{For all measures, calculate the prcc for each parameter}
\usage{
calculate_prcc_for_all_measures(MEASURES, COEFFPARAMCOL, COEFFDATA,
  LHCRESULTFILE)
}
\arguments{
\item{MEASURES}{Simulation output responses}

\item{COEFFPARAMCOL}{Results for the current simulation parameter}

\item{COEFFDATA}{Coefficient data object being created}

\item{LHCRESULTFILE}{Complete simulation results for all parameter sets}
}
\value{
Updated set of parameter correlation coefficient results
}
\description{
For all measures, calculate the prcc for each parameter
}
