% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_analysis.R
\name{calculate_prccs_all_parameters}
\alias{calculate_prccs_all_parameters}
\title{Calculate PRCC values for all parameter-measure pairs}
\usage{
calculate_prccs_all_parameters(PARAMETERS, LHCRESULTFILE, MEASURES)
}
\arguments{
\item{PARAMETERS}{Simulation parameters}

\item{LHCRESULTFILE}{Summary statistics for all LHC parameter sets}

\item{MEASURES}{Simulation output responses}
}
\value{
Correlation coefficients for all pairings
}
\description{
Calculate PRCC values for all parameter-measure pairs
}
