% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_utilities.R
\name{calculate_weights_for_ensemble_model}
\alias{calculate_weights_for_ensemble_model}
\title{Internal function to calculate the weights for all emulators in the ensemble}
\usage{
calculate_weights_for_ensemble_model(all_model_predictions, emulator_test_data,
  measures, emulator_types, num_of_generations = 8e+05)
}
\arguments{
\item{all_model_predictions}{Set of test set predictions obtained for all
emulators in the ensemble}

\item{emulator_test_data}{Data on which the ensemble performance will be
assessed}

\item{measures}{Simulation responses the model should predict}

\item{emulator_types}{Machine learning techniques being employed}

\item{num_of_generations}{Number of generations for which the neural
network that is generating the weights should attempt to converge within}
}
\value{
weights to use for each emulator in the ensemble
}
\description{
Internal function to calculate the weights for all emulators in the ensemble
}
