% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulation_utilities.R
\name{check_acceptable_model_type}
\alias{check_acceptable_model_type}
\title{Internal function to check whether acceptable models have been specified}
\usage{
check_acceptable_model_type(model_type)
}
\arguments{
\item{model_type}{Model type to be generated}
}
\value{
True or False, whether the model is acceptable
}
\description{
Internal function to check whether acceptable models have been specified
}
\keyword{internal}
