% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_boolean}
\alias{check_boolean}
\title{Check that an argument that should be a boolean has been specified correctly}
\usage{
check_boolean(argument, preCheckSuccess, argName)
}
\arguments{
\item{argument}{Value of the argument to check}

\item{preCheckSuccess}{Current status of pre-execution checks}

\item{argName}{Name of the argument, for inclusion in the error message}
}
\value{
Boolean stating the current status of the pre-execution checks, or FALSE if this check fails
}
\description{
Check that an argument that should be a boolean has been specified correctly
}
