% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_column_ranges}
\alias{check_column_ranges}
\title{For aleatory analysis, checks the analysis start and end columns are sensible}
\usage{
check_column_ranges(arguments, filepath, resultfile)
}
\arguments{
\item{arguments}{List of the arguments provided to the called function}

\item{filepath}{Evaluated filepath argument}

\item{resultfile}{Name of the result file of columns to check}
}
\value{
Boolean stating the current status of the pre-execution checks,
or FALSE if this check fails
}
\description{
For aleatory analysis, checks the analysis start and end columns are sensible
}
