% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_consistency_result_type}
\alias{check_consistency_result_type}
\title{Check that the user has declared a-test results either in an R object or in a file name}
\usage{
check_consistency_result_type(arguments, preCheckSuccess)
}
\arguments{
\item{arguments}{List of the arguments provided to the called function}

\item{preCheckSuccess}{Current status of pre-execution checks}
}
\value{
Boolean stating the current status of the pre-execution checks, or FALSE if this check fails
}
\description{
Check that the user has declared a-test results either in an R object or in a file name
}
