% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_double_value_in_range}
\alias{check_double_value_in_range}
\title{Check that a double argument is within a specified range}
\usage{
check_double_value_in_range(argument, argument_name, range_min, range_max)
}
\arguments{
\item{argument}{Value of the argument to check}

\item{argument_name}{Name of the argument, for inclusion in the error message}

\item{range_min}{Minimum of the range}

\item{range_max}{Maximum of the range}
}
\value{
Boolean stating the current status of the pre-execution checks, or FALSE if this check fails
}
\description{
Check that a double argument is within a specified range
}
