% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_efast_sampling_netlogo_args}
\alias{check_efast_sampling_netlogo_args}
\title{Pre-execution checks to perform before the spartan efast samplng technique
is executed for a netlogo simulation}
\usage{
check_efast_sampling_netlogo_args(arguments)
}
\arguments{
\item{arguments}{List of the arguments provided to the called function}
}
\value{
Boolean stating the status of the pre-execution checks
}
\description{
Pre-execution checks to perform before the spartan efast samplng technique
is executed for a netlogo simulation
}
