% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_file_exist}
\alias{check_file_exist}
\title{Checks for the existence of a file}
\usage{
check_file_exist(filepath, filename)
}
\arguments{
\item{filepath}{Directory where the file should be}

\item{filename}{Name of the file}
}
\value{
Boolean showing whether or not the file exists
}
\description{
Checks for the existence of a file
}
