% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Check whether a file exists}
\usage{
check_file_exists(filepath)
}
\arguments{
\item{filepath}{File path to check}
}
\value{
Boolean stating whether this file exists or not
}
\description{
Check whether a file exists
}
