% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{check_file_extension}
\alias{check_file_extension}
\title{Check the file extension of a file and return it}
\usage{
check_file_extension(filename)
}
\description{
Check the file extension of a file and return it
}
\keyword{internal}
