% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_filepath_exists}
\alias{check_filepath_exists}
\title{Check that the filepath of required data or output exists}
\usage{
check_filepath_exists(arguments, preCheckSuccess)
}
\arguments{
\item{arguments}{List of the arguments provided to the called function}

\item{preCheckSuccess}{Current status of pre-execution checks}
}
\value{
Boolean stating the current status of the pre-execution checks, or FALSE if this check fails
}
\description{
Check that the filepath of required data or output exists
}
