% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_function_dependent_paramvals}
\alias{check_function_dependent_paramvals}
\title{Call the correct paramvals check for the calling function, as netlogo & robustness differ}
\usage{
check_function_dependent_paramvals(input_arguments, argument_name)
}
\arguments{
\item{input_arguments}{List of the arguments provided to the called function}

\item{argument_name}{Null in this case, but keeps auto-called functions consistent}
}
\value{
Boolean stating the status of the pre-execution checks
}
\description{
Call the correct paramvals check for the calling function, as netlogo & robustness differ
}
