% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_getATestResult_Input}
\alias{check_getATestResult_Input}
\title{Pre-execution checks to perform before the spartan get aTest results method
within the consistency analysistechnique is executed. Checks all input}
\usage{
check_getATestResult_Input(arguments)
}
\arguments{
\item{arguments}{List of the arguments provided to the called function}
}
\value{
Boolean stating the status of the pre-execution checks
}
\description{
Pre-execution checks to perform before the spartan get aTest results method
within the consistency analysistechnique is executed. Checks all input
}
