% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_global_param_sampling_args}
\alias{check_global_param_sampling_args}
\title{Checks the input values for global parameter sampling techniques}
\usage{
check_global_param_sampling_args(input_arguments, argument_name)
}
\arguments{
\item{input_arguments}{List of the arguments provided to the called function}

\item{argument_name}{Null in this case, but keeps auto-called functions consistent}
}
\value{
Boolean stating the status of the pre-execution checks
}
\description{
Checks the input values for global parameter sampling techniques
}
