% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_graph_output_type}
\alias{check_graph_output_type}
\title{Check the requested graph types are correct (PDF, PNG, TIFF, BMP)}
\usage{
check_graph_output_type(input_arguments, argument_name)
}
\arguments{
\item{input_arguments}{List of the arguments provided to the called function}

\item{argument_name}{Null in this case, but keeps auto-called functions consistent}
}
\value{
Boolean stating the status of the pre-execution checks
}
\description{
Check the requested graph types are correct (PDF, PNG, TIFF, BMP)
}
