% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_input_args}
\alias{check_input_args}
\title{Wrapper function called by all spartan methods to check input pre-execution}
\usage{
check_input_args(argNames, arguments)
}
\arguments{
\item{argNames}{Expected argument names for the calling function}

\item{arguments}{User input to the called function}
}
\value{
Boolean stating whether the input checks pass or fail
}
\description{
Wrapper function called by all spartan methods to check input pre-execution
}
