% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_lengths_parameters_ranges}
\alias{check_lengths_parameters_ranges}
\title{Check that the lengths of the parameters, minimum values, and maximum values, are equal}
\usage{
check_lengths_parameters_ranges(arguments)
}
\arguments{
\item{arguments}{List of the arguments provided to the called function}
}
\value{
Boolean stating the current status of the pre-execution checks, or FALSE if this check fails
}
\description{
Check that the lengths of the parameters, minimum values, and maximum values, are equal
}
