% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_list_all_integers}
\alias{check_list_all_integers}
\title{Check that all objects of a list are integers}
\usage{
check_list_all_integers(argument, preCheckSuccess, argName)
}
\arguments{
\item{argument}{Value of the argument to check}

\item{preCheckSuccess}{Current status of pre-execution checks}

\item{argName}{Name of the argument, for inclusion in the error message}
}
\value{
Boolean stating the current status of the pre-execution checks, or FALSE if this check fails
}
\description{
Check that all objects of a list are integers
}
