% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_nested_filepaths}
\alias{check_nested_filepaths}
\title{Check that result filepaths under the root directory exist}
\usage{
check_nested_filepaths(file_root, sub_dirs)
}
\arguments{
\item{file_root}{Root directory of the data to analyse}

\item{sub_dirs}{List of the subdirectories that should be under the root}
}
\value{
Boolean stating the current status of the pre-execution checks, or FALSE if this check fails
}
\description{
Check that result filepaths under the root directory exist
}
