% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_netlogo_parameters_and_values}
\alias{check_netlogo_parameters_and_values}
\title{Checks the netlogo parameters and values are formatted correctly}
\usage{
check_netlogo_parameters_and_values(arguments, argument_name)
}
\arguments{
\item{arguments}{List of the arguments provided to the called function}

\item{argument_name}{Here for consistency in auto-function calling, but not used}
}
\value{
Boolean stating the current status of the pre-execution checks,
or FALSE if this check fails
}
\description{
Checks the netlogo parameters and values are formatted correctly
}
