% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_package_installed}
\alias{check_package_installed}
\title{Check that a required package has been installed}
\usage{
check_package_installed(packageName)
}
\arguments{
\item{packageName}{Name of the package to check for}
}
\value{
Boolean stating whether the package is installed or not
}
\description{
Check that a required package has been installed
}
