% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_paramvals_length_equals_parameter_length}
\alias{check_paramvals_length_equals_parameter_length}
\title{Where used in robustness analysis, check that the length of PARAMVALS equals}
\usage{
check_paramvals_length_equals_parameter_length(arguments)
}
\arguments{
\item{arguments}{List of the arguments provided to the called function}
}
\value{
Boolean stating the current status of the pre-execution checks,
or FALSE if this check fails
}
\description{
Where used in robustness analysis, check that the length of PARAMVALS equals
}
