% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_robustness_parameter_and_ranges_lengths}
\alias{check_robustness_parameter_and_ranges_lengths}
\title{Where used, checks that PARAMETERS, PMIN, PMAX, PINC, and BASELINE are all
the same length}
\usage{
check_robustness_parameter_and_ranges_lengths(arguments)
}
\arguments{
\item{arguments}{List of the arguments provided to the called function}
}
\value{
Boolean stating the current status of the pre-execution checks,
or FALSE if this check fails
}
\description{
Where used, checks that PARAMETERS, PMIN, PMAX, PINC, and BASELINE are all
the same length
}
