% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_robustness_paramvals_contains_baseline}
\alias{check_robustness_paramvals_contains_baseline}
\title{Checks that the parameter values specified in PARAMVALS contain the BASELINE}
\usage{
check_robustness_paramvals_contains_baseline(arguments)
}
\arguments{
\item{arguments}{List of the arguments provided to the called function}
}
\value{
Boolean stating the current status of the pre-execution checks,
or FALSE if this check fails
}
\description{
The paramvals need to contain the BASELINE value else no behaviours can be
compared using the robustness analysis approach in Technique 2
}
