% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_robustness_range_or_values}
\alias{check_robustness_range_or_values}
\title{For robustness, check whether using PMIN/PMAX/PINC entry or PARAMVALS}
\usage{
check_robustness_range_or_values(arguments, argument_name)
}
\arguments{
\item{arguments}{List of the arguments provided to the called function}

\item{argument_name}{Name of the argument being checked.}
}
\value{
Boolean stating the current status of the pre-execution checks,
or FALSE if this check fails
}
\description{
For robustness, check whether using PMIN/PMAX/PINC entry or PARAMVALS
}
