% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_robustness_sampling_args}
\alias{check_robustness_sampling_args}
\title{Pre-execution checks to perform before the spartan robustness samplng
technique is executed. Checks all parameter input}
\usage{
check_robustness_sampling_args(arguments, argument_name)
}
\arguments{
\item{arguments}{List of the arguments provided to the called function}

\item{argument_name}{Null in this case, but keeps auto-called functions consistent}
}
\value{
Boolean stating the status of the pre-execution checks
}
\description{
Pre-execution checks to perform before the spartan robustness samplng
technique is executed. Checks all parameter input
}
