% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netlogo_sampling_utilities.R
\name{close_and_write_netlogo_file}
\alias{close_and_write_netlogo_file}
\title{Close the current netlogo sample file and write out}
\usage{
close_and_write_netlogo_file(xml, FILENAME)
}
\arguments{
\item{xml}{Object of the XML file being constructed}

\item{FILENAME}{Name of the file to write out}
}
\description{
Close the current netlogo sample file and write out
}
