% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness_analysis.R
\name{compare_all_values_of_parameter_to_baseline}
\alias{compare_all_values_of_parameter_to_baseline}
\title{For one parameter, compare responses for all values with those at baseline}
\usage{
compare_all_values_of_parameter_to_baseline(parameter_value_list,
  parameters, parameter_index, baseline, result, exp_params,
  baseline_result, measures, all_atest_scores)
}
\arguments{
\item{parameter_value_list}{List of values for this parameter}

\item{parameters}{Simulation parameters}

\item{parameter_index}{Index of the current parameter being analysed}

\item{baseline}{Baseline values of all parameters}

\item{result}{Simulation results, read in from CSV file, to subset}

\item{exp_params}{Current parameter set being analysed}

\item{baseline_result}{Simulation result under baseline conditions}

\item{measures}{Simulation output measures}

\item{all_atest_scores}{Store of output of this analysis, to append to}
}
\value{
all_atest_scores with resuls for this parameter appended
}
\description{
For one parameter, compare responses for all values with those at baseline
}
