% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network_utilities.R
\name{create_neural_network}
\alias{create_neural_network}
\title{Create neural network emulator, using neuralnet package}
\usage{
create_neural_network(formula, training_set, fold_number, normalised_data,
  network_structure, number_of_generations)
}
\arguments{
\item{formula}{Parameters and measures formula to use in creating the
network}

\item{training_set}{Training data on which to train the network}

\item{fold_number}{Number of the fold being created in k-fold cross
validation}

\item{normalised_data}{Normalised version of the training data}

\item{network_structure}{Hidden layers structure to use to create the
network}

\item{number_of_generations}{Number of generations to train the
network within in the hope of convergence, else training terminates}
}
\value{
Neural network created
}
\description{
Create neural network emulator, using neuralnet package
}
