% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network_utilities.R
\name{createAndEvaluateFolds}
\alias{createAndEvaluateFolds}
\title{Create and evaluate folds within k-fold cross validation}
\usage{
createAndEvaluateFolds(fold_size, test_fold_start, test_fold_end, dataset,
  network_struc, formula, parameters, measures, algorithm_settings)
}
\arguments{
\item{fold_size}{Number of rows of the dataset that form each fold}

\item{test_fold_start}{Starting position of the training or test fold in the
dataset}

\item{test_fold_end}{End position of the training or test fold in the
dataset}

\item{dataset}{Dataset for which the training and test folds are being
developed}

\item{network_struc}{Network structure for which the folds are being assessed}

\item{formula}{Parameters and measures formula to use in creating the
network}

\item{parameters}{Names of the parameters that form the input nodes of
the neural network}

\item{measures}{Names of the simulation responses that form the output
node of the neural network}

\item{algorithm_settings}{Object output from the function
emulation_algorithm_settings, containing the settings of the machine
learning algorithms to use in emulation creation. In this case, the
settings parameter we are interested in is number of generations}
}
\value{
MSE errors for all network structures
}
\description{
Create and evaluate folds within k-fold cross validation
}
