% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network_utilities.R
\name{createtest_fold}
\alias{createtest_fold}
\title{Create test data fold for k-fold cross validation}
\usage{
createtest_fold(fold, number_of_folds, test_fold_start, test_fold_end, dataset)
}
\arguments{
\item{fold}{Number of the fold being examined}

\item{number_of_folds}{Total number of folds}

\item{test_fold_start}{Position at which the test fold starts in the dataset}

\item{test_fold_end}{Position at which the test fold ends in the dataset}

\item{dataset}{Testing data}
}
\value{
Testing data for this fold of k-fold cross validation
}
\description{
Create test data fold for k-fold cross validation
}
