% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network_utilities.R
\name{createTrainingFold}
\alias{createTrainingFold}
\title{Create training data fold for k-fold cross validation}
\usage{
createTrainingFold(fold, number_of_folds, training_fold_start,
  training_fold_end, dataset)
}
\arguments{
\item{fold}{Number of the fold being examined}

\item{number_of_folds}{Total number of folds}

\item{training_fold_start}{Position at which the training fold starts in the
dataset}

\item{training_fold_end}{Position at which the training fold ends in the
dataset}

\item{dataset}{Training data}
}
\value{
Training data for this fold of k-fold cross validation
}
\description{
Create training data fold for k-fold cross validation
}
