% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{dataset_precheck}
\alias{dataset_precheck}
\title{Before partitioning data, removes any columns where the value is all equal, or all NA}
\usage{
dataset_precheck(dataset, parameters, measures)
}
\arguments{
\item{dataset}{Dataset to be used in emulation development}

\item{parameters}{Simulation parameters being assessed}

\item{measures}{Simulation measures being assessed}
}
\value{
List containing the amended dataset, parameters, and measures, or if no modification, NULL
}
\description{
Before partitioning data, removes any columns where the value is all equal, or all NA
}
