% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{denormalise_dataset}
\alias{denormalise_dataset}
\title{Rescale normalised data back to it's original scale}
\usage{
denormalise_dataset(normalised_data, scaled_mins, scaled_maxes)
}
\arguments{
\item{normalised_data}{Dataset that has been normalised}

\item{scaled_mins}{The minimum value used for each parameter in generating
the latin-hypercube sample}

\item{scaled_maxes}{The maximum value used for each parameter in generating
the latin-hypercube sample}
}
\description{
Rescale normalised data back to it's original scale
}
\keyword{internal}
