% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network_utilities.R
\name{determine_optimal_neural_network_structure}
\alias{determine_optimal_neural_network_structure}
\title{Determine the optimal hidden layer structure from those provided}
\usage{
determine_optimal_neural_network_structure(dataset, parameters, measures,
  algorithm_settings)
}
\arguments{
\item{dataset}{Data on which the neural network is being trained}

\item{parameters}{Names of the parameters that form the input nodes of
the neural network}

\item{measures}{Names of the simulation responses that form the output
node of the neural network}

\item{algorithm_settings}{Object output from the function
emulation_algorithm_settings, containing the settings of the machine
learning algorithms to use in emulation creation. In this case, the
settings parameter we are interested in is number of generations}
}
\value{
Optimal network structure from those provided for assessment
}
\description{
Determine the optimal hidden layer structure from those provided
}
