% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast_utilities.R
\name{efast_cvmethod}
\alias{efast_cvmethod}
\title{Internal function used in calculating the partitions of variance in eFAST}
\usage{
efast_cvmethod(si, range_si, sti, range_sti, OUTMEASURE, NUMPARAMS, NUMCURVES,
  NUMOUTMEASURES)
}
\arguments{
\item{si}{Amount of variance partitioned to each parameter}

\item{range_si}{Range of the Si values across all curves}

\item{sti}{Amount of variance accounted for by non-linear effects with other
parameters, for each parameter}

\item{range_sti}{Range of the STi values across all curves}

\item{OUTMEASURE}{Simulation output measure being analysed}

\item{NUMPARAMS}{Number of parameters in the simulation}

\item{NUMCURVES}{Number of resample curves being examined}

\item{NUMOUTMEASURES}{Number of simulation output measures in total}
}
\description{
Internal function used in calculating the partitions of variance in eFAST
}
\keyword{internal}
