% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast_plotting.R
\name{efast_graph_Results}
\alias{efast_graph_Results}
\title{Plot the parition of variance in a simulation response for each measure}
\usage{
efast_graph_Results(RESULTS_FILE_PATH, PARAMETERS, si, sti, errors_si,
  errors_sti, MEASURES, TIMEPOINT, TIMEPOINTSCALE,
  output_types = c("pdf"))
}
\arguments{
\item{RESULTS_FILE_PATH}{Where the eFAST results were saved to}

\item{PARAMETERS}{Simulation parameters being explored}

\item{si}{Vector of Si values calculated in eFAST for all parameters}

\item{sti}{Vector of STi values calculated in eFAST for all parameters}

\item{errors_si}{Vector of confidence intervals for Si values for all
parameters}

\item{errors_sti}{Vector of confidence intervals for STi values for all
parameters}

\item{MEASURES}{Simulation output measures}

\item{TIMEPOINT}{Timepoint being analysed}

\item{TIMEPOINTSCALE}{Scale in which the timepoints are measures}

\item{output_types}{Files types of graph to produce (pdf,png,bmp etc)}
}
\description{
Plot the parition of variance in a simulation response for each measure
}
