% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast_analysis_netlogo.R
\name{efast_netlogo_run_Analysis}
\alias{efast_netlogo_run_Analysis}
\title{Deprecated: Use \code{efast_run_Analysis}}
\usage{
efast_netlogo_run_Analysis(FILEPATH, MEASURES, PARAMETERS, NUMCURVES,
  NUMSAMPLES, OUTPUTMEASURES_TO_TTEST, TTEST_CONF_INT, GRAPH_FLAG,
  EFASTRESULTFILENAME, TIMEPOINTS, TIMEPOINTSCALE)
}
\arguments{
\item{FILEPATH}{Directory where the simulation runs can be found, in folders
or in CSV file format}

\item{MEASURES}{Array containing the names of the output measures which
are used to analyse the simulation}

\item{PARAMETERS}{Array containing the names of the parameters of which
parameter samples have been generated}

\item{NUMCURVES}{The number of 'resamples' to perform (see eFAST
documentation) - recommend using at least 3}

\item{NUMSAMPLES}{The number of parameter subsets that were generated in
the eFAST design}

\item{OUTPUTMEASURES_TO_TTEST}{Which measures in the range should be tested
to see if the result is statistically significant.  To do all, and if
there were 3 measures, this would be set to 1:3}

\item{TTEST_CONF_INT}{The level of significance to use for the T-Test
(e.g. 0.95)}

\item{GRAPH_FLAG}{Whether graphs should be produced summarising the output
- should be TRUE or FALSE}

\item{EFASTRESULTFILENAME}{File name under which the full eFAST analysis
should be stored. This will contain the partitioning of variance for each
parameter.}

\item{TIMEPOINTS}{mplemented so this method can be used when analysing
multiple simulation timepoints. If only analysing one timepoint, this
should be set to NULL. If not, this should be an array of timepoints,
e.g. c(12,36,48,60)}

\item{TIMEPOINTSCALE}{Sets the scale of the timepoints being analysed,
e.g. "Hours"}
}
\description{
Deprecated: Use \code{efast_run_Analysis}
}
