% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast_utilities.R
\name{efast_parameterdist}
\alias{efast_parameterdist}
\title{Internal function used in calculating the partitions of variance in eFAST}
\usage{
efast_parameterdist(PARAMVALS, PMAX, PMIN, NUMSAMPLES, NUMPARAMS)
}
\arguments{
\item{PARAMVALS}{Values under which parameters have been examined}

\item{PMAX}{Max sample values of all parameters}

\item{PMIN}{Min sample values of all parameters}

\item{NUMSAMPLES}{Number of samples to take from the sinusoidal curves}

\item{NUMPARAMS}{Number of parameters being analysed}
}
\description{
Internal function used in calculating the partitions of variance in eFAST
}
\keyword{internal}
