% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast_analysis.R
\name{efast_run_Analysis}
\alias{efast_run_Analysis}
\title{Runs the eFAST Analysis for the pre-generated summary file}
\usage{
efast_run_Analysis(FILEPATH, MEASURES, PARAMETERS, NUMCURVES, NUMSAMPLES,
  OUTPUTMEASURES_TO_TTEST, TTEST_CONF_INT, GRAPH_FLAG, EFASTRESULTFILENAME,
  TIMEPOINTS = NULL, TIMEPOINTSCALE = NULL, GRAPHTIME = NULL,
  check_done = FALSE, output_types = c("pdf"),
  csv_file_curve_summary = TRUE, write_csv_file_out = TRUE,
  CURVE_SUMMARY = NULL)
}
\arguments{
\item{FILEPATH}{Directory where the simulation runs can be found, in folders
or in CSV file format}

\item{MEASURES}{Array containing the names of the output measures which
are used to analyse the simulation}

\item{PARAMETERS}{Array containing the names of the parameters of which
parameter samples have been generated}

\item{NUMCURVES}{The number of 'resamples' to perform (see eFAST
documentation) - recommend using at least 3}

\item{NUMSAMPLES}{The number of parameter subsets that were generated in
the eFAST design}

\item{OUTPUTMEASURES_TO_TTEST}{Which measures in the range should be tested
to see if the result is statistically significant.  To do all, and if
there were 3 measures, this would be set to 1:3}

\item{TTEST_CONF_INT}{The level of significance to use for the T-Test
(e.g. 0.95)}

\item{GRAPH_FLAG}{Whether graphs should be produced summarising the output
- should be TRUE or FALSE}

\item{EFASTRESULTFILENAME}{File name under which the full eFAST analysis
should be stored. This will contain the partitioning of variance for each
parameter.}

\item{TIMEPOINTS}{mplemented so this method can be used when analysing
multiple simulation timepoints. If only analysing one timepoint, this
should be set to NULL. If not, this should be an array of timepoints,
e.g. c(12,36,48,60)}

\item{TIMEPOINTSCALE}{Sets the scale of the timepoints being analysed,
e.g. "Hours"}

\item{GRAPHTIME}{Value never needs stating, used internally to produce
graphs when processing multiple timepoints}

\item{check_done}{If using multiple timepoints, whether data entry has been
checked}

\item{output_types}{Files types of graph to produce (pdf,png,bmp etc)}

\item{csv_file_curve_summary}{Whether curve summaries are provided in
multiple CSV files (as in traditional spartan), or as an R object (as in
spartanDB)}

\item{write_csv_file_out}{Whether the analysis should output a CSV file
(TRUE) or return an R object (FALSE)}

\item{CURVE_SUMMARY}{If providing an R object (csv_file_curve_summary=FALSE),
this is the R object containing those summaries}
}
\value{
Either NULL if a CSV file is returned, or an R object containing the
results of this analysis
}
\description{
Produces a file summarising the analysis; partitioning the variance between
parameters and providing relevant statistics. These include, for each
parameter of interest, first-order sensitivity index (Si), total-order
sensitivity index (STi), complementary parameters sensitivity index (SCi),
and relevant p-values and error bar data calculated using a two-sample
t-test and standard error respectively. For a more detailed examination of
 this analysis, see the references in the R Journal paper. For ease of
 representation, the method also produces a graph showing this data for
 each simulation output measure. These graphs and summaries can be produced
 for multiple timepoints.
}
