% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast_utilities.R
\name{efast_sd}
\alias{efast_sd}
\title{Internal function used in calculating the partitions of variance in eFAST}
\usage{
efast_sd(RESULTSARRAY, omi, MI, OUTMEASURES, NUMPARAMS, NUMCURVES)
}
\description{
Internal function used in calculating the partitions of variance in eFAST
}
\keyword{internal}
