% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast_utilities.R
\name{efast_setfreq}
\alias{efast_setfreq}
\title{Internal function used in calculating the partitions of variance in eFAST}
\usage{
efast_setfreq(NUMPARAMS, omci_max, PARAMNUM)
}
\description{
Internal function used in calculating the partitions of variance in eFAST
}
\keyword{internal}
