% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast_utilities.R
\name{efast_ttest}
\alias{efast_ttest}
\title{Internal function used in calculating the partitions of variance in eFAST}
\usage{
efast_ttest(si, range_si, sti, range_sti, OUTPUTMEASURES_TO_TTEST,
  NUMPARAMS, NUMCURVES, TTEST_CONF_INT)
}
\description{
Internal function used in calculating the partitions of variance in eFAST
}
\keyword{internal}
