% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{emulated_lhc_values}
\alias{emulated_lhc_values}
\title{Latin-hypercube value set use to demonstrate emulated sensitivity analysis}
\format{A list with 500 rows (one per parameter set) and six columns}
\usage{
data(emulated_lhc_values)
}
\description{
A dataset containing values for the six parameters that control the
simulation detailed in the case study section of the vignette. These
parameters are defined the accompanying publications referred to in
the vignette
}
\details{
\itemize{
  \item stableBindProbability. Parameter values between 0 and 100
  \item chemokineExpressionThreshold. Parameter values between 0 and 1
  \item initialChemokineExpressionValue. Parameter values between 0.1 and 0.5
  \item maxChemokineExpressionValue. Parameter values between 0.015 and 0.08
  \item maxProbabilityOfAdhesion. Parameter values between 0 and 1
  \item adhesionFactorExpressionSlope. Parameter values between 0.25 and 5
}
}
\keyword{datasets}
