% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_utilities.R
\name{ensemble_abc_wrapper}
\alias{ensemble_abc_wrapper}
\title{Wrapper to allow EasyABC functions to run using Ensemble}
\usage{
ensemble_abc_wrapper(x)
}
\arguments{
\item{x}{Set of parameter values generated by an EasyABC method}
}
\value{
Ensemble prediction using parameter set in x
}
\description{
Provides a means of running the ensemble within the EasyABC methods. This
method should be stated as the "model" argument of EasyABC methods such as
ABC_sequential. Note that as arguments cannot be passed to the model
function, these must be created before calling any EasyABC method (see
code description for \code{create_abc_settings_object}. The created object
most contain the simulation parameters, output measures, an ensemble object
to be run, declared as built_ensemble, and whether or not the prior
(and generated predictions) should be normalised (normalise). Should this
object not exist an error message will be produced.
}
