% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{execute_checks}
\alias{execute_checks}
\title{Executes the list of check functions compiled for the calling function}
\usage{
execute_checks(check_methods_to_call, input_arguments, function_args)
}
\arguments{
\item{check_methods_to_call}{List of check functions that should be called}

\item{input_arguments}{The user's input to a function}

\item{function_args}{The expected names of the arguments, provided by
calling function}
}
\value{
Boolean stating whether the input checks pass or fail
}
\description{
This returns TRUE if all arguments pass, or FALSE when the first test fails
}
