% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast_analysis.R
\name{format_efast_result_for_output}
\alias{format_efast_result_for_output}
\title{Joins the various results objects into an output ready format}
\usage{
format_efast_result_for_output(efast_analysis_stats, t_tests,
  OUTPUTMEASURES_TO_TTEST, MEASURES, PARAMETERS)
}
\arguments{
\item{efast_analysis_stats}{Sensitivity indices}

\item{t_tests}{T-Test scores for output measures}

\item{OUTPUTMEASURES_TO_TTEST}{Which measures in the range should be tested
to see if the result is statistically significant.  To do all, and if
there were 3 measures, this would be set to 1:3}

\item{MEASURES}{Array containing the names of the output measures which
are used to analyse the simulation}

\item{PARAMETERS}{Array containing the names of the parameters of which
parameter samples have been generated}
}
\value{
formatted results for output to csv file
}
\description{
Joins the various results objects into an output ready format
}
