% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aleatory_analysis.R
\name{generate_a_test_results_header}
\alias{generate_a_test_results_header}
\title{Generates the CSV file header for the A-Test results file}
\usage{
generate_a_test_results_header(pre_measure_info, measures)
}
\arguments{
\item{pre_measure_info}{Any header info to put before measure names}

\item{measures}{The simulation output responses}
}
\value{
Header object for CSV file
}
\description{
Generates the CSV file header for the A-Test results file
}
