% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast_sampling.R
\name{generate_efast_parameter_sets}
\alias{generate_efast_parameter_sets}
\title{Use the eFAST approach to generate parameter sets}
\usage{
generate_efast_parameter_sets(FILEPATH, NUMCURVES, NUMSAMPLES, PARAMETERS,
  PMIN, PMAX)
}
\arguments{
\item{FILEPATH}{Directory where the parameter samples should be output to}

\item{NUMCURVES}{The number of 'resamples' to perform (see eFAST
documentation) - recommend using at least 3}

\item{NUMSAMPLES}{The number of parameter subsets to generate - should be at
least 65 for eFAST}

\item{PARAMETERS}{Array containing the names of the parameters of which
parameter samples will be generated.  For eFAST, remember to add a
parameter named 'Dummy'}

\item{PMIN}{Array containing the minimum value that should be used for each
parameter and the dummy. Sets a lower bound on sampling space}

\item{PMAX}{Array containing the maximum value that should be used for each
parameter and the dummy. Sets an upper bound on sampling space}
}
\description{
Use the eFAST approach to generate parameter sets
}
