% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aleatory_analysis.R
\name{generate_headers_for_atest_file}
\alias{generate_headers_for_atest_file}
\title{Generates headers for the A-Test summary CSV and R Object}
\usage{
generate_headers_for_atest_file(measures)
}
\arguments{
\item{measures}{Simulation output responses}
}
\value{
header list for A-Test summary object
}
\description{
Generates headers for the A-Test summary CSV and R Object
}
