% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{generate_list_of_checks}
\alias{generate_list_of_checks}
\title{Defines which functions to call to check an input argument.}
\usage{
generate_list_of_checks(argNames)
}
\arguments{
\item{argNames}{Expected argument names for the calling function}
}
\value{
List of check functions that should be called to check input
}
\description{
This creates a list of check functions to call for the input arguments
for a spartan function. This is using the expected argument names,
provided by the calling function
}
