% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulation_utilities.R
\name{generate_model_formula}
\alias{generate_model_formula}
\title{Internal function to generate the formula for training the models}
\usage{
generate_model_formula(parameters, measures)
}
\arguments{
\item{parameters}{Parameters on which the model will take as input}

\item{measures}{Simulation responses the model should predict}
}
\value{
formula object for use in model creation
}
\description{
Internal function to generate the formula for training the models
}
\keyword{internal}
