% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness_sampling.R
\name{generate_parameter_table}
\alias{generate_parameter_table}
\title{Takes the value list and generates the sample that is output to csv file}
\usage{
generate_parameter_table(PARAMETERS, BASELINE, PARAMOFINT, val_list)
}
\arguments{
\item{PARAMETERS}{Array containing the names of the parameters of which parameter samples will be generated}

\item{BASELINE}{Array containing the values assigned to each of these parameters in the calibrated baseline}

\item{PARAMOFINT}{Number of the parameter for which samples are currently being built}

\item{val_list}{List of parameter values to output for each parameter, to be combined with the baseline value of the complementary set}
}
\value{
Parameter table ready for output to CSV file
}
\description{
Takes the value list and generates the sample that is output to csv file
}
