% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_analysis.R
\name{generate_prcc_results_header}
\alias{generate_prcc_results_header}
\title{Generates the CSV file header for the prcc results file}
\usage{
generate_prcc_results_header(measures)
}
\arguments{
\item{measures}{The simulation output responses}
}
\value{
Header object for CSV file
}
\description{
Generates the CSV file header for the prcc results file
}
