% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulation_utilities.R
\name{generate_predictions_from_emulator}
\alias{generate_predictions_from_emulator}
\title{Internal function to generate predictions from an emulator}
\usage{
generate_predictions_from_emulator(emulator, parameters, measures,
  data_to_predict)
}
\arguments{
\item{parameters}{Parameters on which the model will take as input}

\item{measures}{Simulation responses the model should predict}

\item{data_to_predict}{Unseen values for the parameters for which the
measures should be predicted. Has been pre-processed such that it is
in the correct scale for the emulation.}

\item{emulation}{The emulation object to use to make the predictions}
}
\value{
Predictions generated for this unseen data, in the same scale
as the normalised data used to train the emulation.
}
\description{
Internal function to generate predictions from an emulator
}
\keyword{internal}
