% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{get_argument_correct_case}
\alias{get_argument_correct_case}
\title{Tries upper and lower case names for input arguments}
\usage{
get_argument_correct_case(arguments, argName)
}
\arguments{
\item{arguments}{List of the arguments provided to the called function}

\item{argName}{Name of the argument to return}
}
\value{
The evaluated input value of this argument
}
\description{
The method checking relies on the user calling the variables the right name.
If they don't, an error is produced. This helper function checks they haven't
just used lower case arguments, in which case the lower case is checked
}
