% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{get_correct_file_path_for_function}
\alias{get_correct_file_path_for_function}
\title{Gets the correct filepath for the column range input checker}
\usage{
get_correct_file_path_for_function(arguments)
}
\arguments{
\item{arguments}{Input arguments object passed from the checked
function}
}
\value{
Filepath to the results file to check
}
\description{
check_column_ranges takes a filepath containing a sample results file,
and this changes across functions. This takes the function name
and returns the correct argument names for passing to the function
}
