% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_plotting.R
\name{get_correlation_stat_data}
\alias{get_correlation_stat_data}
\title{Retrieve the correlation stats data to plot}
\usage{
get_correlation_stat_data(FILEPATH, corcoeffs_outputfile,
  corcoeffs_outputobject)
}
\arguments{
\item{FILEPATH}{Root path of the analysis being performed (this may be deleted)}

\item{corcoeffs_outputfile}{Path to the the file containing this data, or NULL}

\item{corcoeffs_outputobject}{Name of the R object containing this data, or NULL}
}
\value{
Data for processing in the plots
}
\description{
Retrieve the correlation stats data to plot
}
