% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{get_file_and_object_argument_names}
\alias{get_file_and_object_argument_names}
\title{Gets the correct file and R object argument names for the input checker}
\usage{
get_file_and_object_argument_names(input_arguments)
}
\arguments{
\item{input_arguments}{Input arguments object passed from the checked
function}
}
\value{
List containg argument names for file and R object
}
\description{
check_consistency_result_type takes the argument name of a file or R
object, and this changes across functions. This takes the function name
and returns the correct argument names for passing to the function
}
