% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_plotting.R
\name{get_lhc_summary_data}
\alias{get_lhc_summary_data}
\title{Retrieve the simulation summary data to plot}
\usage{
get_lhc_summary_data(FILEPATH, lhcsummary_filename, lhcsummary_object)
}
\arguments{
\item{FILEPATH}{Root path of the analysis being performed (this may be deleted)}

\item{lhcsummary_filename}{Path to the the file containing this data, or NULL}

\item{lhcsummary_object}{Name of the R object containing this data, or NULL}
}
\value{
Data for processing in the plots
}
\description{
Retrieve the simulation summary data to plot
}
