% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aleatory_analysis.R
\name{get_max_and_median_atest_scores}
\alias{get_max_and_median_atest_scores}
\title{Return the max and median A-Test score for all measures for a sample size}
\usage{
get_max_and_median_atest_scores(sample_processing, measures,
  sample_size_result)
}
\arguments{
\item{sample_processing}{Sample size being analysed}

\item{measures}{Simulation output responses}

\item{sample_size_result}{all processed simulation responses}
}
\value{
Summary of median and max A-Test scores for this sample size
}
\description{
Return the max and median A-Test score for all measures for a sample size
}
