% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{get_median_results_for_all_measures}
\alias{get_median_results_for_all_measures}
\title{For a model result, calculate the medians of the desired measures}
\usage{
get_median_results_for_all_measures(model_result, measures)
}
\arguments{
\item{model_result}{Simulation results}

\item{measures}{Measures to summarise}
}
\value{
Median summary statistics for all measures
}
\description{
For a model result, calculate the medians of the desired measures
}
