% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{getMediansSubset}
\alias{getMediansSubset}
\title{Function used to calculate the median results for those run under a set of
parameter samples}
\usage{
getMediansSubset(FILEPATH, NUMRUNSPERSAMPLE, measures, resultfilename,
  altfilename = NULL, outputfilecolstart, outputfilecolend)
}
\arguments{
\item{FILEPATH}{Where the simulation results can be found}

\item{NUMRUNSPERSAMPLE}{Number of times the simulation has been run for a
parameter set}

\item{outputfilecolstart}{Number of the column of the CSV file where results
commence}

\item{outputfilecolend}{Number of the column of the CSV file where results
end}

\item{MEASURES}{Names of Simulation output responses}

\item{RESULTFILENAME}{Name of the simulation result file}

\item{ALTFILENAME}{Where necessary, name of the alternative result file}
}
\value{
Median Simulation responses under the parameter set in the
result file
}
\description{
Internal function used to calculate the median set where a simulation has
been run a number of times for a given parameter set. Used in all of
Techniques 1-4 detailed in the R Journal
}
\keyword{internal}
