% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aleatory_analysis.R
\name{graph_sample_size_results}
\alias{graph_sample_size_results}
\title{Graph the A-Test results for a sample size}
\usage{
graph_sample_size_results(FILEPATH, GRAPHNAME, SAMPLESIZE, SIZE_RESULTS,
  MEASURES, LARGEDIFFINDICATOR)
}
\arguments{
\item{FILEPATH}{Where the results can be found}

\item{GRAPHNAME}{If multiple timepoints, holds the time so this can be appended to graph name}

\item{SAMPLESIZE}{Current sample size being analysed}

\item{SIZE_RESULTS}{Simulation results for this sample size}

\item{MEASURES}{Simulation output responses}

\item{LARGEDIFFINDICATOR}{Value either side of 0.5 that indicates a large difference}
}
\description{
Graph the A-Test results for a sample size
}
