% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{import_model_result}
\alias{import_model_result}
\title{Import a model result from either a CSV or XML file}
\usage{
import_model_result(fileaddress, resultfilename, altfilename,
  outputfilecolstart = NULL, outputfilecolend = NULL)
}
\arguments{
\item{fileaddress}{Directory where the file is}

\item{resultfilename}{Name of the results file}

\item{altfilename}{If no results in resultfile, can read an alternative}

\item{outputfilecolstart}{Start column of output in CSV file}

\item{outputfilecolend}{End column of output in CSV file}
}
\value{
Results for this simulation run
}
\description{
Import a model result from either a CSV or XML file
}
