% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netlogo_sampling_utilities.R
\name{initialise_netlogo_xml_file}
\alias{initialise_netlogo_xml_file}
\title{Initialises the Netlogo setup file for this experiment}
\usage{
initialise_netlogo_xml_file(EXPERIMENT, SAMPLE, EXPERIMENT_REPETITIONS,
  RUN_METRICS_EVERYSTEP, NETLOGO_SETUP_FUNCTION, NETLOGO_RUN_FUNCTION,
  MEASURES)
}
\arguments{
\item{EXPERIMENT}{Test for the experiment tag of the file (e.g. LHC, eFAST, etc)}

\item{SAMPLE}{Number of the sample being processed}

\item{EXPERIMENT_REPETITIONS}{The number of times Netlogo should repeat the experiment for each set of parameter values.}

\item{RUN_METRICS_EVERYSTEP}{Boolean stating whether Netlogo should produce output for each timestep.}

\item{NETLOGO_SETUP_FUNCTION}{The name of the function in Netlogo that sets up the simulation. Commonly is named setup.}

\item{NETLOGO_RUN_FUNCTION}{The name of the function in Netlogo that starts the simulation. Commonly named go.}

\item{MEASURES}{Array containing the names of the Netlogo output measures which are used to analyse the simulation.}
}
\value{
Initialised XML file object
}
\description{
Initialises the Netlogo setup file for this experiment
}
