% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_plotting.R
\name{iterate_results_for_plotting}
\alias{iterate_results_for_plotting}
\title{Function used by walk to create plots for all measure and response pairs}
\usage{
iterate_results_for_plotting(measure, data_to_plot, parameters,
  coefficients, output_types, graph_time, filepath)
}
\arguments{
\item{measure}{Simulation measure being examined}

\item{data_to_plot}{Data to plot for all parameters for this measure}

\item{parameters}{Simulation parameters being analysed}

\item{coefficients}{Calculated coefficient values for this analysis}

\item{output_types}{File types of graphs to produce}

\item{graph_time}{If multiple timepoints, the timepoint being plotted}

\item{filepath}{The directory where the graphs are to be saved to}
}
\description{
Function used by walk to create plots for all measure and response pairs
}
