% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{join_strings}
\alias{join_strings}
\title{Join strings and separate by specified character}
\usage{
join_strings(string_list, sepr)
}
\arguments{
\item{string_list}{List of text elements to combine}

\item{sepr}{Character to separate elements by}
}
\value{
Combined list as string separated by specified character
}
\description{
Join strings and separate by specified character
}
\keyword{internal}
