% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{join_strings_nospace}
\alias{join_strings_nospace}
\title{Join strings and separate with no space}
\usage{
join_strings_nospace(string_list)
}
\arguments{
\item{string_list}{List of text elements to combine}
}
\value{
Combined list as string separated by no spaces
}
\description{
Join strings and separate with no space
}
\keyword{internal}
