% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network_utilities.R
\name{kfoldCrossValidation}
\alias{kfoldCrossValidation}
\title{Perform k-fold cross validation for assessing neural network structure
performance}
\usage{
kfoldCrossValidation(dataset, parameters, measures, algorithm_settings)
}
\arguments{
\item{dataset}{Data on which the neural network is being trained and
assessed}

\item{parameters}{Names of the parameters that form the input nodes of
the neural network}

\item{measures}{Names of the simulation responses that form the output
node of the neural network}

\item{algorithm_settings}{Object output from the function
emulation_algorithm_settings, containing the settings of the machine
learning algorithms to use in emulation creation. In this case, the
settings parameter we are interested in is number of generations}
}
\value{
Mean Squared errors for all potential structures
}
\description{
Perform k-fold cross validation for assessing neural network structure
performance
}
