% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_prcc_utilities.R
\name{lhc_calculatePRCCForMultipleTimepoints}
\alias{lhc_calculatePRCCForMultipleTimepoints}
\title{Calculates the PRCC for each parameter at each timepoint, storeing PRCC
and P-Value in two different files to make the plot function easier}
\usage{
lhc_calculatePRCCForMultipleTimepoints(FILEPATH, CORCOEFFSOUTPUTFILE,
  TIMEPOINTS, MEASURES)
}
\arguments{
\item{FILEPATH}{Directory containing all PRCC containing result files}

\item{CORCOEFFSOUTPUTFILE}{Name of the file containing the PRCCS for
one timepoint. Assume that the timepoint is appended on to the end (i.e.
results_12.csv for hour 12 - the timepoint is appended by spartan so
only specify results.csv}

\item{TIMEPOINTS}{Simulation timepoints to analyse}

\item{MEASURES}{Simulation output measures being analysed, in a vector}
}
\description{
Calculates the PRCC for each parameter at each timepoint, storeing PRCC
and P-Value in two different files to make the plot function easier
}
