% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_prcc_utilities.R
\name{lhc_constructcoeff_dataset}
\alias{lhc_constructcoeff_dataset}
\title{Utility function used to create data structure for coefficient output}
\usage{
lhc_constructcoeff_dataset(LHCRESULTFILE, PARAMNAME, PARAMETERS)
}
\description{
Utility function used to create data structure for coefficient output
}
\keyword{internal}
