% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_plotting.R
\name{lhc_countSignificantParametersOverTime}
\alias{lhc_countSignificantParametersOverTime}
\title{Count number of significant (p<0.01) parameters over a timecourse}
\usage{
lhc_countSignificantParametersOverTime(FILEPATH, MEASURES, TIMEPOINTS)
}
\arguments{
\item{FILEPATH}{Where pre-processed P-Value dataset is found}

\item{MEASURES}{Simulation output responses vector}

\item{TIMEPOINTS}{Timepoints to analyse}
}
\description{
Count number of significant (p<0.01) parameters over a timecourse
}
