% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_prcc_utilities.R
\name{lhc_generate_netlogo_PRCoEffs}
\alias{lhc_generate_netlogo_PRCoEffs}
\title{Deprecated. Use \code{lhc_generatePRCoEffs} instead}
\usage{
lhc_generate_netlogo_PRCoEffs(FILEPATH, PARAMETERS, MEASURES,
  LHCSUMMARYFILENAME, CORCOEFFSOUTPUTFILE)
}
\arguments{
\item{FILEPATH}{Directory where the simulation runs of single CSV file can
be found}

\item{PARAMETERS}{Array containing the names of the parameters of which
parameter samples will be generated}

\item{MEASURES}{Array containing the names of the output measures which are
used to analyse the simulation}

\item{LHCSUMMARYFILENAME}{Name of the LHC Summary file to be generated.
Contains each parameter set alongside the result gained when the simulation
was run under that criteria.}

\item{CORCOEFFSOUTPUTFILE}{Name of the generated CSV file generated}
}
\description{
Deprecated. Use \code{lhc_generatePRCoEffs} instead
}
