% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_analysis.R
\name{lhc_generatePRCoEffs_db_link}
\alias{lhc_generatePRCoEffs_db_link}
\title{Generate Partial Rank Correlation Coefficients for parameter/response pairs for results in database}
\usage{
lhc_generatePRCoEffs_db_link(db_results, parameters, measures,
  cor_calc_method = c("s"))
}
\arguments{
\item{db_results}{Set of experiment results from the DB}

\item{parameters}{Simulation parameters}

\item{measures}{Simulation measures}

\item{cor_calc_method}{Way to calculate the correlation coefficient: Pearson's
("p"), Spearman's ("s"), and Kendall's ("k"). Default is p}
}
\description{
For each parameter, and each simulation output measure, calculates the
Partial Rank Correlation Coefficient between the parameter value and the
simulation results, giving a statistical measurement of any effect that
is present. In this case, results are mined from a database, as created by
the spartanDB package, and the statistics returned for adding back to the DB.
}
