% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_analysis.R
\name{lhc_generatePRCoEffs_overTime}
\alias{lhc_generatePRCoEffs_overTime}
\title{Pre-process analysis settings if multiple timepoints are being considered}
\usage{
lhc_generatePRCoEffs_overTime(FILEPATH, PARAMETERS, MEASURES,
  LHCSUMMARYFILENAME, CORCOEFFSOUTPUTFILE, TIMEPOINTS, TIMEPOINTSCALE)
}
\arguments{
\item{FILEPATH}{Directory where the simulation runs of single CSV file can
be found}

\item{PARAMETERS}{Array containing the names of the parameters of which
parameter samples will be generated}

\item{MEASURES}{Array containing the names of the output measures which are
used to analyse the simulation}

\item{LHCSUMMARYFILENAME}{Name of the LHC Summary file to be generated.
Contains each parameter set alongside the result gained when the simulation
was run under that criteria.}

\item{CORCOEFFSOUTPUTFILE}{Name of the generated CSV file generated}

\item{TIMEPOINTS}{Implemented so this method can be used when analysing
multiple simulation timepoints. If only analysing one timepoint, this should
be set to NULL. If not, this should be an array of timepoints,
e.g. c(12,36,48,60)}

\item{TIMEPOINTSCALE}{Implemented so this method can be used when analysing
multiple simulation timepoints. Sets the scale of the timepoints being
analysed, e.g. "Hours"}
}
\description{
Pre-process analysis settings if multiple timepoints are being considered
}
