% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_preprocess.R
\name{lhc_generateTimepointFiles}
\alias{lhc_generateTimepointFiles}
\title{Generates spartan-compatible timepoint files if simulation results over
time are in one file}
\usage{
lhc_generateTimepointFiles(FILEPATH, SPARTAN_PARAMETER_FILE,
  RUN_SUMMARY_FILE_NAME, NUMSAMPLES, NUMRUNSPERSAMPLE, TIMEPOINTS)
}
\arguments{
\item{FILEPATH}{Filepath to all LHC results being analysed}

\item{SPARTAN_PARAMETER_FILE}{CSV file generated by spartan, containing
all parameter sets under which the simulation was run}

\item{RUN_SUMMARY_FILE_NAME}{Simulation output file, containing multiple
timepoints}

\item{NUMSAMPLES}{Number of parameter samples generated using the hypercube}

\item{NUMRUNSPERSAMPLE}{Number of replicate runs performed for each sample}

\item{TIMEPOINTS}{Simulation timepoints to extract from the result file}
}
\description{
Generates spartan-compatible timepoint files if simulation results over
time are in one file
}
