% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_plotting.R
\name{lhc_graphMeasuresForParameterChange_from_db}
\alias{lhc_graphMeasuresForParameterChange_from_db}
\title{Generates parameter/measure plot for each pairing in the analysis, from results stored in a database}
\usage{
lhc_graphMeasuresForParameterChange_from_db(db_results, corcoeffs,
  parameters, measures, MEASURE_SCALE, output_directory,
  OUTPUT_TYPE = c("PDF"))
}
\arguments{
\item{db_results}{Results for a specified experiment mined from the database}

\item{corcoeffs}{Correlation coefficients calculated for those results,
held in the databae}

\item{parameters}{Parameters included in this analysis}

\item{measures}{Simulation output measures}

\item{MEASURE_SCALE}{Scale in which each of the output responses is
measured. Used to label plots}

\item{output_directory}{Folder where the graphs should be stored}

\item{OUTPUT_TYPE}{Type of graph to plot. Can be PDF, PNG, TIFF, BMP, etc,
all formats supported by ggplot2}
}
\description{
Produces a graph for each parameter, and each output measure, showing
the simulation output achieved when that parameter was assigned that value.
Eases identification of any non-linear effects. This method uses simulation
results stored in a database by spartanDB
}
