% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_plotting.R
\name{lhc_graphPRCCForMultipleTimepoints}
\alias{lhc_graphPRCCForMultipleTimepoints}
\title{Produce a plot of PRCC values obtained at multiple timepoints}
\usage{
lhc_graphPRCCForMultipleTimepoints(FILEPATH, MEASURES, TIMEPOINTS)
}
\arguments{
\item{FILEPATH}{Directory where the pre-processed PRCC values can be found}

\item{MEASURES}{Simulation response measures}

\item{TIMEPOINTS}{Timepoints being analysed}
}
\description{
Calculates the PRCC for each parameter at each timepoint in the TIMEPOINTS
vector. Unlike the other methods in Spartan, this stores PRCC and P-Value
in 2 different files to make the plot function easier.
}
