% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_plotting.R
\name{lhc_netlogo_graphMeasuresForParameterChange}
\alias{lhc_netlogo_graphMeasuresForParameterChange}
\title{Make the LHC output plot
Deprecated. Use \code{lhc_graphMeasuresForParameterChange} instead}
\usage{
lhc_netlogo_graphMeasuresForParameterChange(FILEPATH, parameters, measures,
  MEASURE_SCALE, corcoeffs_outputfile, lhcsummary_filename, TIMEPOINTS,
  TIMEPOINTSCALE)
}
\arguments{
\item{FILEPATH}{Directory where the simulation runs of single CSV file can
be found}

\item{parameters}{Array containing the names of the parameters of which
parameter samples will be generated}

\item{measures}{Array containing the names of the output measures which are
used to analyse the simulation}

\item{MEASURE_SCALE}{Scale in which each of the output responses is
measured. Used to label plots}

\item{corcoeffs_outputfile}{File produced by spartan containing the Partial
Rank Correlation Coefficients for each parameter/measure pairing}

\item{lhcsummary_filename}{Name of the LHC Summary file to be generated.
Contains each parameter set alongside the result gained when the simulation
was run under that criteria.}

\item{TIMEPOINTS}{Implemented so this method can be used when analysing
multiple simulation timepoints. If only analysing one timepoint, this should
be set to NULL. If not, this should be an array of timepoints,
e.g. c(12,36,48,60)}

\item{TIMEPOINTSCALE}{Implemented so this method can be used when analysing
multiple simulation timepoints. Sets the scale of the timepoints being
analysed, e.g. "Hours"}
}
\description{
Make the LHC output plot
Deprecated. Use \code{lhc_graphMeasuresForParameterChange} instead
}
