% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{make_extension}
\alias{make_extension}
\title{Adds an extension to a filename}
\usage{
make_extension(file, extension)
}
\arguments{
\item{file}{Name of the file being created}

\item{extenstion}{Type of file being created}
}
\value{
Filename with extension
}
\description{
Adds an extension to a filename
}
\keyword{internal}
