% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{make_filename}
\alias{make_filename}
\title{Makes a filename from a list of strings, separated by "_"}
\usage{
make_filename(string_list)
}
\arguments{
\item{string_list}{List of text elements to combine}
}
\value{
Combined list as a string separated by "_"
}
\description{
Makes a filename from a list of strings, separated by "_"
}
\keyword{internal}
