% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_plotting.R
\name{make_graph_title}
\alias{make_graph_title}
\title{Make graph title, sub title, and file name}
\usage{
make_graph_title(filepath, parameter, graph_time, measure, measure_scale,
  corr_stat, timepointscale)
}
\arguments{
\item{filepath}{Directory to output graph to}

\item{parameter}{Current parameter being processed}

\item{graph_time}{Timepoint, if multiple timepoints}

\item{measure}{Current measure being processed}

\item{measure_scale}{Scale of the measure being processed}

\item{corr_stat}{The PRCC for this parameter-measure pair}

\item{timepointscale}{Scale of timepoints, if multiple}
}
\value{
List containing file, title, and subtitle, and axes labels
}
\description{
Make graph title, sub title, and file name
}
