% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_plotting.R
\name{make_lhc_plot}
\alias{make_lhc_plot}
\title{Uses ggplot2 to create a plot for the LHC analysis, for one parameter-measure pair}
\usage{
make_lhc_plot(param_data, parameter, coefficient, measure, graph_time,
  graph_format = c("PDF"), filepath)
}
\arguments{
\item{param_data}{List of parameter value against simulation responses}

\item{parameter}{Name of the parameter being plotted}

\item{coefficient}{Coefficient value for this parameter measure pairing}

\item{measure}{Name of the measure being plotted}

\item{graph_time}{If plotting multiple timepoints, the timepoint being plotted}

\item{graph_format}{Output formats to produce. Defaults to PDF}

\item{filepath}{The directory where the graphs are to be stored}
}
\description{
Uses ggplot2 to create a plot for the LHC analysis, for one parameter-measure pair
}
