% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{make_path}
\alias{make_path}
\title{Combines a list of elements into a filepath separated by "/"}
\usage{
make_path(string_list)
}
\arguments{
\item{string_list}{List of text elements to combine}
}
\value{
Combined list as string separated by "/"
}
\description{
Combines a list of elements into a filepath separated by "/"
}
\keyword{internal}
