% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulation_utilities.R
\name{meanSquaredError}
\alias{meanSquaredError}
\title{Calculate the mean squared error between predictions and observed data}
\usage{
meanSquaredError(model_predictions, observed_data)
}
\arguments{
\item{model_predictions}{Predicted dataset#}

\item{observed_data}{Observed dataset (testing or validation)}
}
\value{
Mean Squared Error between the two sets
}
\description{
Calculate the mean squared error between predictions and observed data
}
\keyword{internal}
