% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{normalise_dataset}
\alias{normalise_dataset}
\title{Normalise a dataset such that all values are between 0 and 1}
\usage{
normalise_dataset(dataset, sample_mins, sample_maxes, parameters)
}
\arguments{
\item{dataset}{LHC Summary file being used in emulator creation}

\item{sample_mins}{The minimum value used for each parameter in generating
the latin-hypercube sample}

\item{sample_maxes}{The maximum value used for each parameter in generating
the latin-hypercube sample}

\item{parameters}{Simulation parameters the emulation will be fed as input}
}
\value{
List of the scaled data and the mimumum/maximum sample values for
each, to aid rescale of the data and any predictions made using it.
}
\description{
Normalise a dataset such that all values are between 0 and 1
}
