% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atest_utilities.R
\name{normaliseATest}
\alias{normaliseATest}
\title{Normalises the A-Test such that it is above 0.5}
\usage{
normaliseATest(result)
}
\arguments{
\item{result}{A-Test score to be normalised}
}
\value{
Normalised A-Test score
}
\description{
Normalises the A-Test such that it is above 0.5
}
