% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atest_utilities.R
\name{num.decimals}
\alias{num.decimals}
\title{Diagnostic function used to determine number of decimal places}
\usage{
num.decimals(x)
}
\arguments{
\item{x}{Numeric value to examine}
}
\value{
Number of decimal places
}
\description{
Diagnostic function used to determine number of decimal places
}
