% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{output_ggplot_graph}
\alias{output_ggplot_graph}
\title{Output a ggplot graph in the requested formats}
\usage{
output_ggplot_graph(GRAPHFILE, OUTPUT_TYPE, output_graph)
}
\arguments{
\item{GRAPHFILE}{Path and name of the file to output}

\item{OUTPUT_TYPE}{List of the output types to produce}

\item{output_graph}{Graph to output}
}
\description{
Output a ggplot graph in the requested formats
}
