% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast_sampling.R
\name{output_param_sets_per_curve}
\alias{output_param_sets_per_curve}
\title{Output the generated parameter sets for each curve}
\usage{
output_param_sets_per_curve(FILEPATH, NUMCURVES, PARAMETERS, PARAMETERVALS)
}
\arguments{
\item{FILEPATH}{Directory where the parameter samples should be output to}

\item{NUMCURVES}{The number of 'resamples' to perform (see eFAST
documentation) - recommend using at least 3}

\item{PARAMETERS}{Array containing the names of the parameters of which
parameter samples will be generated.  For eFAST, remember to add a
parameter named 'Dummy'}

\item{PARAMETERVALS}{The parameter sets generated by the eFAST sampling method}
}
\description{
Output the generated parameter sets for each curve
}
