% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_prcc_utilities.R
\name{pcor.rec}
\alias{pcor.rec}
\title{Internal function used to calculate the Partial Rank Correlation Coefficient}
\usage{
pcor.rec(x, y, z, cor_method = "p", na.rm = TRUE)
}
\description{
Internal function used to calculate the Partial Rank Correlation Coefficient
}
\keyword{internal}
