% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_prcc_utilities.R
\name{pcor.test}
\alias{pcor.test}
\title{Internal function used to calculate the Partial Rank Correlation Coefficient}
\usage{
pcor.test(x, y, z, use = "mat", calc_method = "p", na.rm = TRUE)
}
\description{
Internal function used to calculate the Partial Rank Correlation Coefficient
}
\keyword{internal}
