% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atest_utilities.R
\name{perform_aTest_for_all_sim_measures}
\alias{perform_aTest_for_all_sim_measures}
\title{Performs A-Test to compare all simulation measures}
\usage{
perform_aTest_for_all_sim_measures(PARAMETER_SET, BASELINE_RESULT,
  PARAMETER_SET_RESULT, MEASURES)
}
\arguments{
\item{PARAMETER_SET}{Set of simulation parameters for which a set of runs was performed}

\item{BASELINE_RESULT}{Simulation behaviour under baseline conditions}

\item{PARAMETER_SET_RESULT}{Simulation behaviour under conditions in this parameter set}

\item{MEASURES}{An array containing the names of the simulation output measures to be analysed.

 @export}
}
\description{
for this set, the simulation baseline behaviour, and performs the A-Test
to get a comparison for all simulation measures
}
