% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulation_plotting.R
\name{plot_compare_sim_observed_to_model_prediction}
\alias{plot_compare_sim_observed_to_model_prediction}
\title{Internal function used to create accuracy plots of the emulation against observed data}
\usage{
plot_compare_sim_observed_to_model_prediction(observed, predicted, technique,
  measure, mse, graph_file_name, timepoint = NULL)
}
\arguments{
\item{observed}{Observed dataset (testing or validation)}

\item{predicted}{Predicted dataset}

\item{technique}{The machine learning technique used to develop the emulator}

\item{measure}{The simulation output response being plotted}

\item{mse}{Mean Squared Error between predicted and observed}

\item{graph_file_name}{Name to give the produced PDF plot}

\item{timepoint}{If using multiple timepoints, the timepoint for which the
emulator has been created}
}
\description{
Outputs plot to PDF in the current working directory
}
