% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast_plotting.R
\name{ploteFASTSiFromTimepointFiles}
\alias{ploteFASTSiFromTimepointFiles}
\title{Plot the Si value for all parameters for multiple simulation timepoints}
\usage{
ploteFASTSiFromTimepointFiles(FILEPATH, PARAMETERS, MEASURES,
  EFASTRESULTFILENAME, TIMEPOINTS, TIMEPOINTSCALE)
}
\arguments{
\item{FILEPATH}{Where the eFAST results have been stored}

\item{PARAMETERS}{Names of simulation parameters being explored}

\item{MEASURES}{Names of simulation output responses}

\item{EFASTRESULTFILENAME}{Name of the CSV file output by eFAST Analysis,
containing all the Si and STi values}

\item{TIMEPOINTS}{Timepoints to include in this analysis}

\item{TIMEPOINTSCALE}{Scale in which the timepoints are measured}
}
\description{
Permits easy comparison of when a parameter may become more influential
than others throughout a simulation timecourse
}
