% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{prepare_parameter_value_list}
\alias{prepare_parameter_value_list}
\title{Prepares the parameter value list, as either an interval range or specific
values can be supplied}
\usage{
prepare_parameter_value_list(PMIN, PMAX, PINC, PARAMVALS, PARAM_OF_INT)
}
\arguments{
\item{PMIN}{Vector of the minimum values for each parameter}

\item{PMAX}{Vector of the maximum values for each parameter}

\item{PINC}{Vector of the sample increment value for each parameter}

\item{PARAMVALS}{Vector of the sampling values specified for each parameter,
rather than an incremental sequence}

\item{PARAM_OF_INT}{The current parameter being analysed}
}
\description{
When iterating through parameter values, we need to do a bit of prep, as
there are two ways these values can be specified (increment and specified
in a list). This returns a list of the values for both specifications, so
these can be iterated through without issue
}
\keyword{internal}
