% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netlogo_sampling_utilities.R
\name{process_netlogo_parameter_range_info}
\alias{process_netlogo_parameter_range_info}
\title{Processes netlogo parameter information to generate names of those of interest to this analysis}
\usage{
process_netlogo_parameter_range_info(PARAMETERS, PARAMVALS)
}
\arguments{
\item{PARAMETERS}{Parameter names specified in the R script}

\item{PARAMVALS}{Values of each parameter, either a specific value or range}
}
\value{
List of: names of parameters of interest, minimum value of each
parameter, and maximum value of each parameter
}
\description{
Not all parameters may be being perturbed, although specified in PARAMETERS.
Some have a specific value, rather than a range. This detects the names of
the parameters that are being perturbed, and their mins and maxes
}
