% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulation_plotting.R
\name{produce_accuracy_plots_all_measures}
\alias{produce_accuracy_plots_all_measures}
\title{Internal function used to create accuracy plots of the emulation against
observed data, for all measures}
\usage{
produce_accuracy_plots_all_measures(technique, measures, model_predictions,
  observed_data, output_format = c("pdf"), timepoint = NULL)
}
\arguments{
\item{technique}{The machine learning technique used to develop the emulator}

\item{measures}{All simulation output responses to plot}

\item{model_predictions}{Predicted dataset}

\item{observed_data}{Observed dataset (testing or validation)}

\item{output_format}{File formats in which graphs should be produced}

\item{timepoint}{If using multiple timepoints, the timepoint for which the
emulator has been created}
}
\description{
Outputs plot to PDF in the current working directory
}
