% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulation_plotting.R
\name{produce_accuracy_plots_single_measure}
\alias{produce_accuracy_plots_single_measure}
\title{Internal function used to create accuracy plots of the emulation against observed data}
\usage{
produce_accuracy_plots_single_measure(technique, measure,
  model_predictions, observed_data, output_format = c("pdf"),
  timepoint = NULL)
}
\arguments{
\item{technique}{The machine learning technique used to develop the emulator}

\item{measure}{The simulation output response being plotted}

\item{model_predictions}{Predicted dataset}

\item{observed_data}{Observed dataset (testing or validation)}

\item{output_format}{File formats in which graphs should be produced}

\item{timepoint}{If using multiple timepoints, the timepoint for which the
emulator has been created}
}
\description{
Outputs plot to PDF in the current working directory
}
