% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aleatory_analysis.R
\name{produce_atest_score_summary}
\alias{produce_atest_score_summary}
\title{Generates A-Test score summary for all sample sizes}
\usage{
produce_atest_score_summary(sample_sizes, allSubset_ATest_Scores, measures)
}
\arguments{
\item{sample_sizes}{Sample sizes being assessed}

\item{allSubset_ATest_Scores}{Scores from all subsets for all sample sizes}

\item{measures}{Simulation output responses}
}
\value{
A-Test max and median summary for all sample sizes
}
\description{
Generates A-Test score summary for all sample sizes
}
