% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast_analysis.R
\name{read_all_curve_results}
\alias{read_all_curve_results}
\title{Reads results from each curve into a multi-dimensional array}
\usage{
read_all_curve_results(FILEPATH, GRAPHTIME, NUMCURVES, NUMSAMPLES, MEASURES,
  PARAMETERS)
}
\arguments{
\item{FILEPATH}{Directory where the simulation runs can be found, in folders
or in CSV file format}

\item{GRAPHTIME}{Value never needs stating, used internally to produce
graphs when processing multiple timepoints}

\item{NUMCURVES}{The number of 'resamples' to perform (see eFAST
documentation) - recommend using at least 3}

\item{NUMSAMPLES}{The number of parameter subsets that were generated in
the eFAST design}

\item{MEASURES}{Array containing the names of the output measures which
are used to analyse the simulation}

\item{PARAMETERS}{Array containing the names of the parameters of which
parameter samples have been generated}
}
\description{
Reads results from each curve into a multi-dimensional array
}
