% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{read_from_csv}
\alias{read_from_csv}
\title{To save retyping all options, function to read CSV data}
\usage{
read_from_csv(filepath)
}
\arguments{
\item{filepath}{Path to CSV file to read}
}
\value{
Data from CSV file
}
\description{
To save retyping all options, function to read CSV data
}
