% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{read_model_result_file}
\alias{read_model_result_file}
\title{Reads a model result file, either CSV or XML}
\usage{
read_model_result_file(fileaddress, resultfilename,
  outputfilecolstart = NULL, outputfilecolend = NULL)
}
\arguments{
\item{fileaddress}{Folder where the result file can be found}

\item{resultfilename}{Name of the result file}

\item{outputfilecolstart}{Start column of output in CSV file}

\item{outputfilecolend}{End column of output in CSV file}
}
\value{
Results for this simulation run
}
\description{
Reads a model result file, either CSV or XML
}
