% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aleatory_analysis.R
\name{read_simulation_results}
\alias{read_simulation_results}
\title{Read in the simulation results either from a file, or R object
The existance of these results was checked in pre-execution checks}
\usage{
read_simulation_results(FILEPATH, AA_SIM_RESULTS_FILE,
  AA_SIM_RESULTS_OBJECT)
}
\arguments{
\item{FILEPATH}{Where the results can be found}

\item{AA_SIM_RESULTS_FILE}{If the objects are in a file, the file name}

\item{AA_SIM_RESULTS_OBJECT}{If in an R object, the name of the object}
}
\value{
Simulation results for processing
}
\description{
Read in the simulation results either from a file, or R object
The existance of these results was checked in pre-execution checks
}
