% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aleatory_analysis.R
\name{retrieve_results_for_comparison_result_set}
\alias{retrieve_results_for_comparison_result_set}
\title{Get the first result set, to which all others are compared}
\usage{
retrieve_results_for_comparison_result_set(RESULT, SAMPLESIZE)
}
\arguments{
\item{RESULT}{Simulation results}

\item{SAMPLESIZE}{Current sample size being examined}
}
\value{
Results for set 1 of this sample size
}
\description{
Get the first result set, to which all others are compared
}
