% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulation_utilities.R
\name{rSquared}
\alias{rSquared}
\title{Calculate the R squared statistic for predictions and observed data}
\usage{
rSquared(model_predictions, observed_data)
}
\arguments{
\item{model_predictions}{Predicted dataset}

\item{observed_data}{Observed dataset (testing or validation)}
}
\value{
R squared statistic for the two sets
}
\description{
Calculate the R squared statistic for predictions and observed data
}
\keyword{internal}
