% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_sampling.R
\name{sample_parameter_space}
\alias{sample_parameter_space}
\title{Generate the LHC design for the chosen algorithm}
\usage{
sample_parameter_space(ALGORITHM, NUMSAMPLES, PARAMETERS)
}
\arguments{
\item{ALGORITHM}{Choice of algorithm to use to generate the hypercube.
Can be set to either 'normal' or 'optimum'. Beware optimum can take a
long time to generate an optimised parameter set (more than 24 hours
in some circumstances)}

\item{NUMSAMPLES}{The number of parameter subsets to generate}

\item{PARAMETERS}{Array containing the names of the parameters of which
parameter samples will be generated}
}
\value{
Latin-Hypercube sample
}
\description{
Generate the LHC design for the chosen algorithm
}
