% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network_utilities.R
\name{selectSuitableStructure}
\alias{selectSuitableStructure}
\title{Selects the most suitable neural network structure from the potentials made}
\usage{
selectSuitableStructure(network_errors)
}
\arguments{
\item{network_errors}{MSE obtained for each measure for all neural network
layer structures examined}
}
\value{
Lowest error network structure
}
\description{
The user will provide a list of neural network objects to assess. These
will be assessed on their performance by mean squared error. This method
simply selects the most suitable structure (where MSE is lowest)
}
