% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_analysis.R
\name{summarise_replicate_runs}
\alias{summarise_replicate_runs}
\title{Summarises replicate runs of a parameter set. Used by LHC and eFAST}
\usage{
summarise_replicate_runs(lhc_all_sim_results, PARAMETERS, MEASURES,
  bind_params = TRUE)
}
\arguments{
\item{lhc_all_sim_results}{All sim results for all parameter sets}

\item{PARAMETERS}{Array containing the names of the parameters of which
parameter samples will be generated}

\item{MEASURES}{Array containing the names of the output measures which are
used to analyse the simulation}

\item{bind_params}{Whether to include the parameter values in the set of results
(eFAST doesn't)}
}
\value{
Summary of responses under each parameter set
}
\description{
Summarises replicate runs of a parameter set. Used by LHC and eFAST
}
