% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{tutorial_consistency_set}
\alias{tutorial_consistency_set}
\title{Example dataset showing the structure for consistency analysis data}
\format{A list with 9060 rows and four columns}
\usage{
data(tutorial_consistency_set)
}
\description{
This dataset contains exemplar results from a consistency analysis, that
shown in the tutorial. This data can be read in and processed using the
functions detailed in the vignette. The important thing to obtain from this
object is the structure the data is expected to be in
}
\details{
\itemize{
  \item SampleSize. Number of simulation executions in each subset
  \item Set. The number of the subset this result belongs to
  \item Velocity. The median velocity observed in simulation executions in
  this set
  \item Displacement. The median displacement observed in simulation
  executions in this set
}
}
\keyword{datasets}
