% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_utilities.R
\name{weight_emulator_predictions_by_ensemble}
\alias{weight_emulator_predictions_by_ensemble}
\title{Internal function to weight emulator predictions by that calculated for the
ensemble}
\usage{
weight_emulator_predictions_by_ensemble(model_weights, all_model_predictions,
  measures, num_generations = 8e+05)
}
\arguments{
\item{model_weights}{Weights for all emulators in the ensemble, based on
test set performance}

\item{all_model_predictions}{Set of test set predictions obtained for all
emulators in the ensemble}

\item{measures}{Simulation responses the model should predict}

\item{num_generations}{Number of generations for which the neural
network that is generating the weights should attempt to converge within}
}
\value{
predictions generated by the ensemble
}
\description{
Internal function to weight emulator predictions by that calculated for the
ensemble
}
