% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartan_utilities.R
\name{write_data_to_csv}
\alias{write_data_to_csv}
\title{Shortcut function for writing data to CSV file}
\usage{
write_data_to_csv(outputData, outputFile)
}
\arguments{
\item{outputData}{Data to write to CSV file}

\item{outputFile}{Name of the output file}
}
\description{
Shortcut function for writing data to CSV file
}
