% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_analysis.R
\name{calculate_prcc_for_all_measures}
\alias{calculate_prcc_for_all_measures}
\title{For all measures, calculate the prcc for each parameter}
\usage{
calculate_prcc_for_all_measures(MEASURES, COEFFPARAMCOL, COEFFDATA,
  LHCRESULTFILE, cor_calc_method = c("s"), prcc_method = "mat")
}
\arguments{
\item{MEASURES}{Simulation output responses}

\item{COEFFPARAMCOL}{Results for the current simulation parameter}

\item{COEFFDATA}{Coefficient data object being created}

\item{LHCRESULTFILE}{Complete simulation results for all parameter sets}

\item{cor_calc_method}{Way to calculate the correlation coefficient: Pearson's
("p"), Spearman's ("s"), and Kendall's ("k"). Default is p}

\item{prcc_method}{Method to calculate the partial correlation coefficient, either
variance-covariance matrix ("mat") or recursive formula ("rec"). Default mat}
}
\value{
Updated set of parameter correlation coefficient results
}
\description{
For all measures, calculate the prcc for each parameter
}
