% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_analysis.R
\name{calculate_prccs_all_parameters}
\alias{calculate_prccs_all_parameters}
\title{Calculate PRCC values for all parameter-measure pairs}
\usage{
calculate_prccs_all_parameters(PARAMETERS, LHCRESULTFILE, MEASURES,
  cor_calc_method = c("s"))
}
\arguments{
\item{PARAMETERS}{Simulation parameters}

\item{LHCRESULTFILE}{Summary statistics for all LHC parameter sets}

\item{MEASURES}{Simulation output responses}

\item{cor_calc_method}{Way to calculate the correlation coefficient: Pearson's
("p"), Spearman's ("s"), and Kendall's ("k"). Default is p}
}
\value{
Correlation coefficients for all pairings
}
\description{
Calculate PRCC values for all parameter-measure pairs
}
