% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_argument_positive_int}
\alias{check_argument_positive_int}
\title{Check that an argument that should be a positive integer has been specified correctly}
\usage{
check_argument_positive_int(arguments, argument_name)
}
\arguments{
\item{arguments}{List of the arguments provided to the called function}

\item{argument_name}{Name of the argument, for inclusion in the error message}
}
\value{
Boolean stating the current status of the pre-execution checks, or FALSE if this check fails
}
\description{
Check that an argument that should be a positive integer has been specified correctly
}
