% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_filepath_exists}
\alias{check_filepath_exists}
\title{Check that the filepath of required data or output exists}
\usage{
check_filepath_exists(arguments, argument_name)
}
\arguments{
\item{arguments}{Input arguments}

\item{argument_name}{Name of the argument, as provided by function}
}
\value{
Boolean stating the current status of the pre-execution checks, or FALSE if this check fails
}
\description{
Check that the filepath of required data or output exists
}
