% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_list_all_integers}
\alias{check_list_all_integers}
\title{Check that all objects of a list are integers}
\usage{
check_list_all_integers(arguments, argument_name)
}
\arguments{
\item{arguments}{List of the arguments provided to the called function}

\item{argument_name}{Name of the argument, for inclusion in the error message}
}
\value{
Boolean stating the current status of the pre-execution checks, or FALSE if this check fails
}
\description{
Check that all objects of a list are integers
}
