% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_argument_check.R
\name{check_numeric_list_values}
\alias{check_numeric_list_values}
\title{Check that two lists are numeric, and the values of one are less than the other}
\usage{
check_numeric_list_values(arguments, nameSmall, nameLarge)
}
\arguments{
\item{arguments}{List of the arguments provided to the called function}

\item{nameSmall}{Parameter name of the smaller list, for error reporting}

\item{nameLarge}{Parameter name of the larger list, for error reporting}
}
\value{
Boolean stating the current status of the pre-execution checks, or FALSE if this check fails
}
\description{
Check that two lists are numeric, and the values of one are less than the other
}
