% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_utilities.R
\name{graph_Posteriors_All_Parameters}
\alias{graph_Posteriors_All_Parameters}
\title{Graph posterior distributions generated for all parameters, to PDF file}
\usage{
graph_Posteriors_All_Parameters(abc_resultset, parameters, sampleMins,
  sampleMaxes)
}
\arguments{
\item{abc_resultset}{Result object obtained from the EasyABC package}

\item{parameters}{Array containing the names of the parameters for which
posterior distributions will be generated}

\item{sampleMins}{Minimum value of the range over which each parameter was
explored using ABC}

\item{sampleMaxes}{Maximum value of the range over which each parameter was
explored using ABC}
}
\description{
We provide a means of plotting the produced posterior distribution for all
parameters for which the value is being explored. Output to PDF in the
working directory
}
