% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhc_plotting.R
\name{make_lhc_plot}
\alias{make_lhc_plot}
\title{Make the LHC output plot}
\usage{
make_lhc_plot(data_to_plot, titles)
}
\arguments{
\item{data_to_plot}{Parameter and measure pair data}

\item{titles}{Object containing graph title and subtitle}
}
\value{
Created graph object
}
\description{
Make the LHC output plot
}
