% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast_analysis.R
\name{read_all_curve_results}
\alias{read_all_curve_results}
\title{Reads results from each curve into a multi-dimensional array}
\usage{
read_all_curve_results(FILEPATH, GRAPHTIME, NUMCURVES, NUMSAMPLES,
  MEASURES, PARAMETERS)
}
\arguments{
\item{FILEPATH}{Filepath where simulation results can be found}

\item{GRAPHTIME}{Timepoint graph is being produced}

\item{NUMCURVES}{Number of resample curves used in this analysis}

\item{NUMSAMPLES}{Number of samples taken from each curve}

\item{MEASURES}{Simulation output measures of interest}

\item{PARAMETERS}{Simulation parameters of interest}
}
\value{
R matrix containing the curve summaries in the format spartan requires
}
\description{
Reads results from each curve into a multi-dimensional array
}
