% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_generation.R, R/ensemble_utilities.R
\name{use_ensemble_to_generate_predictions}
\alias{use_ensemble_to_generate_predictions}
\alias{use_ensemble_to_generate_predictions}
\title{Use an ensemble to generate predictions for an unseen data set}
\usage{
use_ensemble_to_generate_predictions(generated_ensemble, data_to_predict,
  parameters, measures, normalise_values = FALSE, normalise_result = FALSE)

use_ensemble_to_generate_predictions(generated_ensemble, data_to_predict,
  parameters, measures, normalise_values = FALSE, normalise_result = FALSE)
}
\arguments{
\item{generated_ensemble}{Ensemble object to use to make predictions}

\item{data_to_predict}{Unseen parameter set for which values are to be
generated}

\item{parameters}{Names of the simulation parameters}

\item{measures}{Names of the simulation output responses being predicted}

\item{normalise_values}{Whether the unseen parameter sets should be
normalised to be between 0 and 1}

\item{normalise_result}{Whether the generated predictions should be
normalised to be between 0 and 1}

\item{normalise}{Boolean stating whether or not unseen parameter set needs
to be normalised}

\item{generated_ensemble}{Ensemble to use to make predictions}

\item{data_to_predict}{Parameter sets to make predictions from}

\item{parameters}{Simulation parameter names}

\item{measures}{Simulation output response names}
}
\value{
Predictions for all simulation output measures

List of predictions made for specified responses for all
parameter sets
}
\description{
This method takes a set of parameters and generates a prediction of the
simulation responses under those conditions. It should be specified as to
whether the unseen data should be normalied, as this needs to be in the
scale 0-1 to be acceptable to the ensemble.

This takes a set of unseen parameter values and uses an ensemble to make
predictions of the responses that the simulator would generate
}
