\name{Utility: Sample Data}
\alias{AA_Example_ATestMaxAndMedians}
\alias{Curve1_chemoLowerLinearAdjust}
\alias{Curve1_chemoUpperLinearAdjust}
\alias{Curve1_chemoThreshold}
\alias{Curve1_Dummy}
\alias{Curve1_maxVCAMeffectProbabilityCutoff}
\alias{Curve1_Results_Summary}
\alias{Curve1_thresholdBindProbability}
\alias{Curve1_vcamSlope}
\alias{eFAST_Analysis}
\alias{EgSet_LHC_corCoeffs}
\alias{LHC_Summary}
\alias{LHC_Parameters_for_Runs}
\alias{OAT_Example_ATests_chemoLowerLinearAdjust}
\alias{OAT_Medians}
\alias{OAT_Sampling_chemoThreshold_Values}
\alias{OAT_Sampling_maxVCAMeffectProbabilityCutoff_Values}
\title{Utility: Sample Data}
\description{Sample Data sets to view for furthering understanding of output generated by the available methods. These are described in the detailed description of each technique above}
\usage{
data(AA_Example_ATestMaxAndMedians)
data(Curve1_chemoLowerLinearAdjust)
data(Curve1_chemoUpperLinearAdjust)
data(Curve1_chemoThreshold)
data(Curve1_Dummy)
data(Curve1_maxVCAMeffectProbabilityCutoff)
data(Curve1_Results_Summary)
data(Curve1_thresholdBindProbability)
data(Curve1_vcamSlope)
data(eFAST_Analysis)
data(EgSet_LHC_corCoeffs)
data(LHC_Parameters_for_Runs)
data(LHC_Summary)
data(OAT_Example_ATests_chemoLowerLinearAdjust)
data(OAT_Medians)
data(OAT_Sampling_chemoThreshold_Values)
data(OAT_Sampling_maxVCAMeffectProbabilityCutoff_Values)
}
\keyword{datasets}
